Make the server extension work
The server only accepts url safe base64 encoded path.

We use namespaced API calls and NAMESPACE/PROJECT_NAME is
already encoded as NAMESPACE%2FPROJECT_NAME and should be passed as is to the proxy.
See https://docs.gitlab.com/ee/api/#namespaced-path-encoding

The issue is that tornado route matching automatically decodes url.
See https://www.tornadoweb.org/en/stable/_modules/tornado/routing.html#PathMatches
So it replaces %2F with / in the matched path.

To avoid this, the path is encoded in base64 on the client side
and decoded in the server extension.